# fairway360/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fairway360 multi-tenant SaaS API. All data is scoped to the authenticated user's club; club_id is always derived from the session, never from the request.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        "401": { $ref: "#/components/responses/Error" }
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request a password-reset email
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ForgotPasswordInput" }
      responses:
        "200":
          description: Always 200 (does not reveal whether the email exists)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Set a new password using a reset token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetPasswordInput" }
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "400": { $ref: "#/components/responses/Error" }
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        "401": { $ref: "#/components/responses/Error" }

  # ---- Public ---------------------------------------------------------------
...
  /channels/{id}/messages:
    get:
      operationId: getChannelMessages
      tags: [members]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      responses:
        "200":
          description: Channel message history
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChatMessage" }
    post:
      operationId: sendChannelMessage
      tags: [members]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChannelMessageInput" }
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatMessage" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }

  # ---- Escalations ----------------------------------------------------------
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
  schemas:
...
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    ChatMessage:
      type: object
      properties:
        id: { type: string }
        channelId: { type: string }
        senderUserId: { type: string }
        senderName: { type: string }
        senderRole: { type: string }
        aiGenerated: { type: boolean }
        content: { type: string }
        time: { type: string }
        createdAt: { type: string }
      required: [id, channelId, senderName, content, createdAt]
    ChannelMessageInput:
      type: object
      properties:
        content: { type: string, minLength: 1 }
      required: [content]
...
    ProblemDetails:
      type: object
      properties:
        title: { type: string }
        detail: { type: string }
        status: { type: number }
      required: [title]
    Role:
      type: string
      enum: [supervisor, employee, member, super_admin]
    LoginRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]
    ForgotPasswordInput:
      type: object
      properties:
        email: { type: string }
      required: [email]
    ResetPasswordInput:
      type: object
      properties:
        token: { type: string }
        password: { type: string, minLength: 8 }
      required: [token, password]
    AuthUser:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        initials: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        clubName: { type: string }
        clubSlug: { type: string }
      required: [id, name, email, role, clubName]
...