# fairway360/lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fairway360 multi-tenant SaaS API. All data is scoped to the authenticated user's club; club_id is always derived from the session, never from the request.
 * OpenAPI spec version: 0.1.0
 */

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  supervisor: 'supervisor',
  employee: 'employee',
  member: 'member',
  super_admin: 'super_admin',
} as const;